Make a feature test more resilient to change
The feature test `spec/features/show_poms_feature_spec.rb` runs against
data held in NOMIS dev (T3). It was previously quite brittle due to
expecting the POM "Moic Pom" to appear 8th in the list of POMs. It's
inevitable that this will change as team members come and go, and the
list of available POMs changes accordingly.

To make the test more resilient to such external change, I've introduced
a new helper method `row_containing`. Now rather than expecting the
"Moic Pom" to always be in 8th position in the table, we can simply find
the table row which contains that POM's name.