MO-826: Write a test to demonstrate the bug
I've made the spec for MovementOnDateJob more comprehensive to cover the
bug which we're observing.

The error actually occurs in the AllocationHistory model – but the only
route to that functionality is through this job. So I'm treating this as
a more 'high level' test of the movement processing functionality as a
whole.