Fix error when CircleCI updates Firefox version
There was a bug which only revealed itself when CircleCI tried to update the cached Firefox installation to a newer version.

The CircleCI browser-tools orb annoyingly doesn't symlink Firefox into the user's local bin directory. We need it installed in the user's directory because this is what gets cached between CI builds. So in order to achieve this, we do some shuffling and move the Firefox symlink into the local bin directory after installation.