Memoise prison offenders to improve performance
It's not uncommon for us to call `#offenders` on Prison objects more than once.

Previously, this method would would generate the list of offenders afresh with every call. Thankfully our HTTP API layer implements caching, so this didn't cause excess network traffic, but it was still an unnecessary overhead.

Instead, we now memoise the offenders list as an instance variable on the prison object. This makes it very inexpensive to call `#offenders` more than once against the same prison object.