MO-843: default to NPS if resourcing not assessed
This commit changes the behaviour of
`OffenderService.get_community_data` to gracefully handle resourcing
responses from the Community API which are missing the
`enhancedResourcing` field.

If the `enhancedResourcing` field is missing, we handle it the same as
if we received a 404 Not Found response – i.e. we default the service
provider to NPS.

Only if the `enhancedResourcing` field is present and `false` will we
designate an offender as being under CRC rules.