MO-615 remove unused feature flags for ndelius -auto delius import. This is WIP as there were a few quirks while I was removing auto-delius-import.
MO-615
Remove the auto_delius_import_enabled test - it feels redundant but it was also passing falsely - displaying 'This service is not available' which worries me. So it might not be the correct thing to do to remove it!

As suggested by the ticket MO-94, I have removed the show action from CaseInformationController as this screen is apparently unreachable via any route when the flag is disabled. I therefore also deleted the show view along side this and the related tests that used the show view to display a link - so the whole test suite delius_process_data_feature_spec.rb was deleted.