Fix a test which should never have passed
The offender in question should have been recommended a Probation POM – but for some reason, the test was expecting them to be recommended a Prison POM.

To make the test pass, I've changed the offender's tier from B to C. This will change the recommended POM type from Probation POM to Prison POM, making the rest of the test pass.

It's a bit of a mystery how this test ever passed in the first place. Even more of a mystery is why it's suddenly stopped passing. 🤷🏻‍♂️