Allow Google Tag Manager to track error pages
This commit adds a JavaScript snippet to each of the error page views
which pushes details of the error to into the `dataLayer` object – which
is used by Google Analytics to receive custom events from the page.

This makes it possible to begin tracking when and how often these pages
are displayed to end users.