Remove the prison debugging page
The prison debugging page (as opposed to the offender debugging page) was unused within the service.

The upcoming move to the Search API (in MO-806 and MO-852) makes this page redundant because we perform offender filtering at a 'lower level' within the application. Where we previously initialised Offender objects for offenders who are filtered out of the service, this will no longer be the case. The filtering happens in the API client rather than passing those offenders into the rest of the application.

This page is safe to remove because it's not used by anybody on the team. In fact, it was actually forgotten about until we re-discovered it. 😅