Cherry pick parole commit to rename field to target_hearing_date
Original commit message: "Updating references to Parole Review Date to
Target Hearing Date"

Alias on model was used to keep old parole_records table as it is for
now, since we'll be moving to a new table in the future. This way
references to that field can be updated without messing with a table
that will soon be replaced.

Not all changes were brought across, so more massaging needs to be
performed.

MO-1414