New non-functional "Upcoming handovers" page added, old page sidelined
It looks and feels like the prototype but does not yet do anything.
A feature flag in the URL params is required to access it. The default
index action redirects to it (again if feature flag set).

The old one is renamed out of the way, and still used if no feature flag
is set.

MO-1115