AllocatedOffender#all uses Enumerator for more efficiency and laziness
It does not store all allocations in memory if it has to; instead, as
it iterates through AllocatedOffenders, it gives the GC a chance to
the memory of used ones.

Secondly, it allows the actual work to be done interleaved with the
API calls; so we get an API call to get all offenders at a prison, then
those are processed, then another API call is made, etc. This does
not change overall job size but puts less burst load on the API.

MO-1114