Remove a bad test
It was hard to maintain, refactoring required the test to be completely
rewritten, and it was testing multiple things at the same time.

Any time the code that it was testing is touched, we can rewrite the
tests from scratch as we have done with the handover date calculations.

MO-1235