Add a link to 'show allocations' page to revert responsibility override
NOTE: I'm deleting the feature test because it depends on a NOMIS
offender existing that has not reached the handover date yet; and since
the march of time will always eventually make that offender reach
handover, this test will keep breaking and forcing us to find ANOTHER
offender that is before handover.

I personally believe that such a test is a cause of more hardship than
benefits and am deleting it.

MO-1243