Remove rubocop-performance
Readability is THE MOST IMPORTANT THING in software development,
especially in this code base. "Performance optimizations" often leads to
code that uses less-known or non-obvious constructs, like

```
Array.new(2) { |x| }
 ```

instead of the more easily read and commonly understandable

```
2.times.map { |x| ... }
```

Code performance is not the bottleneck of MPC, it's the API calls.
Disabling this means we are free to write the most easily understandable
code we can.