Fix local failing test
This test was not failing on CI because it was tagged as `local_only`.

It was failing locally due to the fact the `#get_image` method performs 2 http requests (a post and a get) and both were being recorded by VCR as successful, however there was at the same time an “attempt” to stub the GET request with webmock to return 404, but this was never really happening as VCR takes precedence so the recorded response (non-404) was being returned instead, and that made the test assertion fail.

I found in this instance was easier to just stub both requests, given from the first post we only use one attribute `facialImageId` and use it in the second, get, request, and remove the VCR dependency.