MO-1479 New 403 error response in SAR API (#2417)
MPC returns 401 response when a client hits the SAR endpoint with a valid token that is missing the required role - this should be a 403 instead.

The existing 401 error has been split into two:
- Leave `401` response for missing (no bearer) or invalid tokens (invalid being anything that cannot be decoded, or expired tokens).
- New `403` response when token, despite being valid, lacks the appropriate role.

Note this only affects SAR endpoint, which has its own implementation of the `#verify_token` method.
It is not an app-wise change. We might decide to do the same in all other endpoints.

Not to be merged yet, awaiting some details/confirmation from Andrew.