Fallback POM details if allocation history is nil
It seems for some offenders we might not have allocation history, meaning we can't retrieve POM details either (as the `primary_pom_nomis_id` is used for this, and it comes from the allocation).

This prevent some 1-week chaser emails from triggering, as it raises a nil exception.

I'm using the same approach found in  `FollowUpEmailDetails` to have a fallback in case allocation is nil or inactive.

Note: need to confirm copy/workaround with Laura, so not merging until then.