chore: Update dependencies (#2440)
* chore: Update dependencies

Minor bump to some gems to close dependabots.

`json-schema` had a major bump. We had a lock in its version but I've checked and everything seems to work fine with latest version.

There is some work needed to be able to update notifications client to latest version. Will be done separately as we need to agree copy changes for uploaded files.

Update `dependabot.yml` to include docker, and to only raise PRs for security versions, and not for version updates. This will keep the repo updated and notify us whenever there are security issues but not flood us with minor, unnecessary versions. Minor bumps can be bundled together and done periodically as part of the general maintenance of the app.

* Revert net-pop change

It seems to break the docker build process.