Tweak ProcessDeliusDataJob audit logs
The `ProcessDeliusDataJob` writes to the `AuditEvent` table the before/after attributes but not just the changed ones, all of them.

This is verbose and difficult to actually see what was changed, which is necessary for a related piece of work MO-1227.

Limit the audit to only the attributes that really changed.