Reduce active job arguments logging
A few of our jobs are for emails, and these emails contain some arguments that are PII and should not be logged.

We can avoid this without removing all logs completely in a job by job basis:
https://edgeapi.rubyonrails.org/classes/ActiveJob/Logging.html#method-i-log_arguments

I've also cleanup a few unused files and enabled the delivery of `urgent_pipeline_to_community` mail as I noticed it was not being delivered because it was missing the deliver method call.