MO-1552 Remove welsh questions
Ticket: https://dsdmoj.atlassian.net/browse/MO-1552

These questions are no longer needed. Cleaned up all of the frontend, and some of the backend but not all of it, as the attribute is still set automatically based on the LDU country when the case information record is created or updated as part of the events, and it's used in the `PrisonPolicies` logic.

We can do the remaining cleanup if necessary once this has been clarified.