MO-1188 Show not found error page
Ticket: https://dsdmoj.atlassian.net/browse/MO-1188

If the user is not signed in and access an inexistent URL, instead of the nice "not found" error page, it was blowing up with a 500 internal server error.

Fixed by improving the handling of exceptions in the `dps_header_footer` method.

As a separate PR (awaiting copy confirmation) we will also have a nice error page (different to the not found) for the auth failure scenario.