MO-1566 Remove community API VCR (#2494)
Ticket: https://dsdmoj.atlassian.net/browse/MO-1566

Removed VCR from the tests that were using it in the offender service spec. These are related to delius community API.

These tests have already counterparts with stubbing so VCR was redundant.

There are a few other tests in this spec using VCR for the Prison API that are way too involved to stub as they perform multiple requests to hydrate the data. These may be considered for removal at a later time.