Fix for bad data throwing error in non-prod envs
Ticket: https://dsdmoj.atlassian.net/browse/MO-1571

In local and staging, a prisoner had a nil `prison_arrival_date` making the
`awaiting_allocation_for` calculation throw an error in the "missing information" page.

If the offender was not in the first page of the pagination, it would throw
an error when ordering the table.

Handle potentially bad data in non-prod envs by taking care of nils and
presenting "N/A" as a fallback in the lists.