MO-1068 Headings and copy tweaks to caseload pages
Ticket: https://dsdmoj.atlassian.net/browse/MO-1068

Refer to Jira ticket for full list of changes, but mainly there was a few headings missing, as well as minor copy changes.

Also added sorting to the POM name column (it was missing the link to perform this sorting but the backend was already prepared to handle this).