Add missing sorting method
There is already a similar method in the `OffenderWithAllocationPresenter` but in some pages it seems we have an `AllocatedOffender` instance instead, and the method call is throwing a few errors in Sentry and might be related to a user support query.

https://ministryofjustice.sentry.io/issues/5858044124/events/6414803091b04603bfad7a173d1e287e/events/?project=5584139

I can't reproduce it locally tho so this is my best guess at what is going on judging by the logs/sentry.