MO-1127 Disallow edit case info for linked cases
Ticket: https://dsdmoj.atlassian.net/browse/MO-1127

This was mostly already done in most places, I only found a place where the link was still there for linked cases.

In case we missed some other place, and also to not allow editing by changing the URL, I added a check to the controller action and some logging, just in case.