MO-912 Update 'missing COM' notification copy
Ticket: https://dsdmoj.atlassian.net/browse/MO-912

As per ticket, new copy implemented (it actually simplifies previous copy/code) in the notifications displayed when someone is in their handover window without a COM and either we have details of the relevant LDU or we don’t.

Refer to ticket for more details.