MO-1664 Enable linking to new MPC (#2578)
* MO-1664 Enable linking to new MPC

Ticket: https://dsdmoj.atlassian.net/browse/MO-1664

With the purpose of slowly be able to swap functionality in the current MPC service, with the new MPC one, some links will start to point to new MPC, depending if the feature is enabled in the corresponding environment.

For now only `staging` will have 1 link swapped (parole cases).

* Change of approach

Instead of having to use a class to call the paths (thus needing to make changes to the views, etc.) we overwrite the paths and check if the feature flag is enabled/disable to produce the same outcome more easily and also we have a more explicit list of paths that are migrated.