MO-1676 Preparing Parole Review DB table for S3
Ticket: https://dsdmoj.atlassian.net/browse/MO-1676

In preparation of the new S3 import mechanism, I'm adding this column to the existing `parole_review_imports` to have traceability of which file in the bucket was used for each import.

This is a non-breaking change with the existing IMAP import mechanism.