MO-1690 Deactivate CNLs from released offenders (#2614)
Ticket: https://dsdmoj.atlassian.net/browse/MO-1690

Task to deactivate CNLs from released offenders, mostly very old
records that were created before we handled OUT movements and events.

Task is a one-off run thus is not scheduled (no cronjob) but can be run
again in the future if needed, so keeping it in the repo for reference.