Ensure deallocation publishes the event (#2627)
Follow-up to PR #2625.

The publishing of the allocation events rely on dirty changes to know which attributes were changed, but these will not play nice when the same allocation instance is modified several times as part of the same transaction.

We were getting no changes and thus not publishing the event.