MO-1714 Remove reliance on CircleCI (part 3) (#2633)
* MO-1714 Remove reliance on CircleCI (part 3)

Ticket: https://dsdmoj.atlassian.net/browse/MO-1714

Final (hopefully) PR related to this work. Remove CircleCI pipeline as all tests are now running on GH actions.

There is still scope to remove some or all of the remaining VCR cassettes and also fix the tests that are tagged as `@flaky` and `@local_only` to also run fine on CI.

* Stub bank holidays http request