Support getting just one POM back (#2642)
Instead of always getting the whole list of POMs for a given prison, and then looping to find the one we wanted, we can pass a query param with the staffId we are looking for and will return (if found) just that one POM.

Refactor code `get_single_pom` to use this approach.

Many test stubs had to be updated sadly, but it is more explicit what we are stubbing.