Get working pattern from POM API response (#2643)
Instead of hardcoding to 0.0 we get it from the POM as we already have this information coming through the API response. This will make more sense soon with the POM onboarding.

Updated PomDetails factory with these attributes (were added to the model in previous PR but factory was not committed back then).