Revert to previous 0.0 default for working pattern (#2648)
We need more information and ask around to decide if we want to use the hours we receive from NOMIS and assign a working pattern derived from it in our service.

Currently we default to 0.0 but given we receive the hours from NOMIS we could improve this. However there is a bit of variance in the data coming from NOMIS, with 35 hours being used in a majority of cases as FT, but others like 37 and even 39 are also used. And some in between.

To be safe and to be able to deploy to production other unrelated stuff, put this back as it was and we will decide later.