MO-1758 Remove POM functionality
Ticket: https://dsdmoj.atlassian.net/browse/MO-1758

Remove POMs that are no longer registered as such in NOMIS but they were left with allocated cases in our service.

These, if any, will show up in a new tab `Attention needed` along with the number of cases they have, so the POM can be deleted and their cases free to allocate again.

Part of the backend for this to work has been done over past PRs so this is mostly just the final frontend/controller work.

It is behind a feature flag so it is safe if a deploy to prod happens.