MO-1844 Store username who added a POM (#2702)
Ticket: https://dsdmoj.atlassian.net/browse/MO-1844

There are 2 ways to add POMs to the service, one is NOMIS, and another through the "Add a POM" in the service. We need to gather stats periodically of how many POMs are added with each method.

This will store the username of the currently signed-in SPO who performed the action in the service.